Create swapTokensStructuredOutputParser.py
This is langchain StructuredOutput parsing tool for mUSD to BTC token swaps on DumpySwap on Mezo Matsnet. 

In this parser there are 4 fields in the ResponseSchema. This extra field is to account for two currencies required in a user prompt to swap. 

It does not need a recipient address and will always extract the same DumpySwap router address.